requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.10.3
flower==1.0.1
humanize==3.11.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
sentry-sdk==1.3.1
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5